+++ src/algorithms/libs/rtklib/rtklib_conversions.h
@@ -28,6 +28,7 @@
  * \{ */
 
 
+class Beidou_Dnav_Almanac;
 class Beidou_Dnav_Ephemeris;
 class Galileo_Almanac;
 class Galileo_Ephemeris;
@@ -104,6 +105,7 @@ eph_t eph_to_rtklib(const Beidou_Dnav_Ephemeris& bei_eph);
 
 alm_t alm_to_rtklib(const Gps_Almanac& gps_alm);
 alm_t alm_to_rtklib(const Galileo_Almanac& gal_alm);
+alm_t alm_to_rtklib(const Beidou_Dnav_Almanac& bei_alm);
 
 /*!
  * \brief Transforms a Glonass_Gnav_Ephemeris to its RTKLIB counterpart